requirements.txt
gensim==4.3.3
keras==3.5.0
tensorflow==2.17.1
scikit-learn==1.6.0
wheel==0.45.1
grpcio
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
grpcio-tools
nltk==3.9.1
Keras-Preprocessing
pytest
kagglehub